+++ frontend/src/app/rootSaga.ts
@@ -1,0 +1,7 @@
+import { all } from "redux-saga/effects";
+import counterSaga from "./features/counter/counterSaga";
+import usersSaga from "./features/users/usersSaga";
+
+export default function* rootSaga() {
+  yield all([counterSaga(), usersSaga()]);
+}
