+++ frontend/src/app/store.ts
@@ -1,14 +1,14 @@
 import { configureStore } from "@reduxjs/toolkit";
 import createSagaMiddleware from "redux-saga";
-import { counterReducer } from "./reducer";
-import rootSaga from "../sagas/rootSaga";
+import { rootReducer } from "./reducer";
+import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
 
 // update the reducers you want to use here
 export const store = configureStore({
   reducer: {
-    counter: counterReducer,
+    counter: rootReducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({ thunk: false }).concat(sagaMiddleware),
