+++ frontend/src/app/types/user.d.ts
