--- frontend/src/sagas/rootSaga.ts
@@ -1,6 +1,0 @@
-import { all } from "redux-saga/effects";
-import counterSaga from "../features/counter/counterSaga";
-
-export default function* rootSaga() {
-  yield all([counterSaga()]);
-}
